YipYap/backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenvpython-geohash 
